Fix decode issue
In base64_decode_block(), step_b & step_c have a bug in that they will cause bits of the next byte to be written even though there are no more bytes in the encoded data. This adds an additional check for more encoded data before writing the additional bits to the next byte.